Full lab1

+++ src/main/java/Man.java
@@ -1,18 +1,18 @@
 public class Man {
-    private int closedKey;
-    public int openKey;
+    private long closedKey;
+    public long openKey;
 
-    public  Man(int P,int g){
+    public  Man(long P,long g){
         closedKey = random(P);
         Mod m = new Mod(g,closedKey,P);
         openKey = m.exp2().intValue();
     }
-    public int Call(Man m,int P){
+    public long Call(Man m,long P){
         Mod m1 = new Mod(m.openKey,closedKey,P);
         return m1.exp2().intValue();
     }
-    private static int random(int max)
+    private static long random(long max)
     {
-        return (int) (Math.random() * ++max);
+        return (long) (Math.random() * ++max);
     }
 }
